+++ packages/schema/src/types.ts
@@ -14,7 +14,7 @@ export * from './_interface/project/sources/videoSource';
 export * from './_interface/project/sources/geoParquetSource';
 export * from './_interface/project/sources/markerSource';
 export * from './_interface/project/sources/wmsTileSource';
-export * from './_interface/project/sources/openeoSource';
+export * from './_interface/project/sources/openeoTileSource';
 
 // Layers
 export * from './_interface/project/layers/heatmapLayer';
@@ -26,7 +26,7 @@ export * from './_interface/project/layers/imageLayer';
 export * from './_interface/project/layers/stacLayer';
 export * from './_interface/project/layers/vectorTileLayer';
 export * from './_interface/project/layers/geoTiffLayer';
-export * from './_interface/project/layers/openeoLayer';
+export * from './_interface/project/layers/openeoTileLayer';
 
 // Processing
 export * from './processing/_generated/exportProcessingSchema';
